requirements.txt
beautifulsoup4==4.12.3
certifi==2024.7.4
charset-normalizer==3.3.2
colorama==0.4.6
crayons==0.4.0
idna==3.6
numpy==1.26.4
pandas==2.2.0
pymongo==4.6.3
pytz==2024.1
requests==2.32.0
schedule==1.2.0
selenium==4.17.2
termcolor==2.4.0
webdriver-manager==3.5.2
python-dotenv
